moved the package to o.a.s.cluster.api
