+++ solr/solrj/src/java/org/apache/solr/common/cloud/ZkStateReader.java
@@ -49,8 +49,6 @@ import org.apache.solr.common.Callable;
 import org.apache.solr.common.SolrCloseable;
 import org.apache.solr.common.SolrException;
 import org.apache.solr.common.SolrException.ErrorCode;
-import org.apache.solr.common.cloud.sdk.ShardReplica;
-import org.apache.solr.common.cloud.sdk.SolrNode;
 import org.apache.solr.common.params.CollectionAdminParams;
 import org.apache.solr.common.params.CoreAdminParams;
 import org.apache.solr.common.util.*;
